auditor-agent-py/requirements.txt
fastapi
uvicorn
google-generativeai
httpx
beautifulsoup4
python-dotenv
pydantic
